// ClientApp/src/core/Models/iso-retrieval-model.ts
export default interface IIsoRetrievalModel{
    asn: string
    city:  String
    continent_code: string
    country: string
    country_area: number
    country_calling_code: string
    country_capital: string
    country_code: string
    country_code_iso3: string
    country_name?: string
    country_population: number
    country_tld: string
    currency: string
    currency_name:  string
    in_eu: boolean
    ip: string
    languages:  string
    latitude:   number
    longitude: number
    org:  string
    postal:  string
    region: string
    region_code: string
    timezone:  string
    utc_offset:  string
    version:  string
}